[Refactor] 설정 페이지 코드 리팩토링
